Revert "Merge pull request #24 from Bhaveshupadhyay/feat/rag-pipeline"
This reverts commit 11de97e254c544d90ae4c2ca8d44c4afd7c212d9, reversing
changes made to d22fdee6420ead4b4901a58296ea870ba4a26f74.

+++ src/backend/models/client.py
@@ -7,7 +7,7 @@ class Requirement(BaseModel):
     vector: List[float]
 
 class ExtractedFacts(BaseModel):
-    client_budget: float
+    client_budget: int
     estimated_cost: float = 0.0
     currency: str = "USD"
     tech_stack: List[str] = Field(default_factory=list)
